backend/requirements.txt
pyodbc==4.0.32
pandas==1.5.3
scikit-learn==1.2.2
transformers==4.27.1
torch==1.13.1
numpy==1.24.3
python-dotenv==1.0.0
Flask==2.1.2
requests==2.26.0
beautifulsoup4==4.10.0
flask-cors==3.0.10
